// apps/admin/src/shared/api/generated/ja-korea/schemas/reportResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */
import type { AssetResponse } from './assetResponse';
import type { ReportResponseReportType } from './reportResponseReportType';

export interface ReportResponse {
  id?: number;
  reportType?: ReportResponseReportType;
  title?: string;
  thumbnail?: AssetResponse;
  attachment?: AssetResponse;
  attachmentFileName?: string;
  downloadCount?: number;
  createdAt?: string;
  updatedAt?: string;
  createdBy?: number;
  version?: number;
}